New simd

+++ src/bh.cpp
@@ -213,9 +213,9 @@ void bh_tree_evaluate(const vector<bh_tree_node>& nodes, vector<int>& sink_bucke
 				const simd_float r1overh1 = r * hinv;// 1
 				const simd_float r2oh2 = r1overh1 * r1overh1;// 1
 				simd_float rinv1_near = -5.0f / 16.0f;
-				rinv1_near = fma(rinv1_near, r2oh2, simd_float(21.0f / 16.0f));// 2
-				rinv1_near = fma(rinv1_near, r2oh2, simd_float(-35.0f / 16.0f));// 2
-				rinv1_near = fma(rinv1_near, r2oh2, simd_float(35.0f / 16.0f));// 2
+				rinv1_near = fmaf(rinv1_near, r2oh2, simd_float(21.0f / 16.0f));// 2
+				rinv1_near = fmaf(rinv1_near, r2oh2, simd_float(-35.0f / 16.0f));// 2
+				rinv1_near = fmaf(rinv1_near, r2oh2, simd_float(35.0f / 16.0f));// 2
 				rinv1_near *= hinv;// 1
 				const auto near_flag = (simd_float(1) - far_flag);// 1
 				rinv1 = far_flag * rinv1_far + near_flag * rinv1_near;// 4
